requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
fonttools==4.59.2
frozenlist==1.7.0
idna==3.10
joblib==1.5.1
kiwisolver==1.4.9
matplotlib==3.10.6
multidict==6.6.3
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.20.1
